rename depth_splatter to just splatter.
