--- _RUN_Depth_Splatting_GUI.bat
@@ -1,2 +1,0 @@
-call venv\scripts\activate.bat
-python depth_splatting_gui.py
