+++ _RUN_Splatting_GUI.bat
@@ -1,0 +1,2 @@
+call venv\scripts\activate.bat
+python splatting_gui.py
