+++ requirements.txt
@@ -20,3 +20,4 @@ Imath
 triton-windows
 moviepy==1.0.3
 pyperclip
+ttkthemes
